pyproject.toml
[project]
name = "language-generate"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "lyuweii", email = "wliu@ihep.ac.cn" }]
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "python-igraph>=0.11.8",
]

[dependency-groups]
dev = [
    "ipython>=9.1.0",
    "pytest>=8.3.5",
]
